requirements.txt
aiohttp==3.9.1
aiosignal==1.3.1
altair==5.2.0
annotated-types==0.6.0
anyio==4.1.0
asttokens==2.2.1
async-timeout==4.0.3
attrs==23.1.0
backcall==0.2.0
backports.functools-lru-cache==1.6.5
backports.zoneinfo==0.2.1
blinker==1.7.0
branca==0.6.0
brotlipy==0.7.0
cachetools==5.3.2
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==2.0.4
click==8.1.7
cloudpickle==3.0.0
colorama==0.4.6
contourpy==1.1.0
cryptography==39.0.1
cycler==0.11.0
dataclasses-json==0.6.3
datascience==0.17.6
debugpy==1.5.1
decorator==5.1.1
distro==1.8.0
exceptiongroup==1.2.0
executing==1.2.0
filelock==3.9.0
folium==0.14.0
fonttools==4.40.0
frozenlist==1.4.1
fsspec==2023.12.2
gitdb==4.0.11
GitPython==3.1.40
greenlet==3.0.2
gymnasium==0.28.1
h11==0.14.0
httpcore==1.0.2
httpx==0.25.2
huggingface-hub==0.19.4
idna==3.4
imageio==2.31.1
importlib-metadata==6.8.0
importlib-resources==6.0.0
ipykernel==6.15.0
ipython==8.12.0
jax-jumpy==1.0.0
jedi==0.18.2
Jinja2==3.1.2
joblib==1.3.1
jsonpatch==1.33
jsonpointer==2.4
jsonschema==4.20.0
jsonschema-specifications==2023.11.2
jupyter_client==8.3.0
jupyter_core==5.3.1
kiwisolver==1.4.4
langchain==0.0.350
langchain-community==0.0.3
langchain-core==0.1.1
langsmith==0.0.71
lazy_loader==0.3
llvmlite==0.41.1
markdown-it-py==3.0.0
MarkupSafe==2.1.1
marshmallow==3.20.1
matplotlib==3.7.2
matplotlib-inline==0.1.6
mdurl==0.1.2
mkl-fft==1.3.6
mkl-random==1.2.2
mkl-service==2.4.0
mpmath==1.2.1
multidict==6.0.4
mypy-extensions==1.0.0
nest-asyncio==1.5.6
networkx==2.8.4
numba==0.58.1
numpy==1.24.3
openai==1.4.0
packaging==23.2
pandas==2.0.3
parso==0.8.3
pickleshare==0.7.5
Pillow==9.4.0
pip==23.1.2
pkgutil_resolve_name==1.3.10
platformdirs==2.5.2
plotly==5.15.0
prompt-toolkit==3.0.39
protobuf==4.25.1
psutil==5.9.0
pure-eval==0.2.2
pyarrow==14.0.1
pycparser==2.21
pydantic==2.5.2
pydantic_core==2.14.5
pydeck==0.8.1b0
pygame==2.1.3
Pygments==2.15.1
pyOpenSSL==23.0.0
pyparsing==3.0.9
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
PyWavelets==1.4.1
pywin32==305.1
PyYAML==6.0.1
pyzmq==25.1.0
referencing==0.32.0
regex==2023.10.3
requests==2.29.0
rich==13.7.0
rpds-py==0.13.2
safetensors==0.4.1
scikit-image==0.21.0
scikit-learn==1.3.0
scipy==1.10.1
setuptools==67.8.0
shap==0.44.0
six==1.16.0
slicer==0.0.7
smmap==5.0.1
sniffio==1.3.0
SQLAlchemy==2.0.23
stack-data==0.6.2
streamlit==1.29.0
sympy==1.11.1
tenacity==8.2.2
threadpoolctl==3.1.0
tifffile==2023.7.10
tokenizers==0.15.0
toml==0.10.2
toolz==0.12.0
torch==2.0.1
torchaudio==2.0.2
torchvision==0.15.2
tornado==6.2
tqdm==4.65.0
traitlets==5.9.0
transformers==4.36.1
typing_extensions==4.6.3
typing-inspect==0.9.0
tzdata==2023.3
tzlocal==5.2
urllib3==1.26.16
validators==0.22.0
watchdog==3.0.0
wcwidth==0.2.6
wheel==0.38.4
win-inet-pton==1.1.0
yarl==1.9.4
zipp==3.16.0
